Review fix – move code around
