+++ ydb/public/sdk/cpp/src/client/impl/internal/scheme_helpers/ya.make
@@ -1,0 +1,12 @@
+LIBRARY()
+
+SRCS(
+    helpers.cpp
+)
+
+PEERDIR(
+    ydb/public/api/protos
+    ydb/public/sdk/cpp/src/client/driver
+)
+
+END()
