+++ ydb/public/sdk/cpp/src/client/ya.make
@@ -23,6 +23,7 @@ RECURSE(
     impl/internal/logger
     impl/internal/make_request
     impl/internal/plain_status
+    impl/internal/scheme_helpers
     impl/internal/thread_pool
     impl/internal/value_helpers
     impl/session
